Remove package activation command restriction
It was mistakenly thought that Atom would require the main module of
packages as part of its loading step. Since this is not the case, there is
no apparent means of permitting custom override commands to activate the
package, unless users manually modify the `package.json` file, which is
certainly not an elegant solution.

All changes based on the aforementioned assumption have been removed.
This should fix override commands being unable to trigger the symbol
search display.